requirements.txt
flask
python-binance==1.0.15
pandas
numpy
